adding auth

+++ controlers/itemController.js
@@ -4,6 +4,7 @@ module.exports.createItem = async (req, res) => {
   const itemObj = req.body;
 
   const item = new Item({
+    owner: req.user_id,
     title: itemObj.title,
     slug: itemObj.slug,
     price: itemObj.price,
@@ -24,12 +25,21 @@ module.exports.createItem = async (req, res) => {
 };
 
 module.exports.deleteItem = async (req, res) => {
-  await Item.deleteOne({ _id: req.params.id });
+  const item = Item.findOne({ _id: req.params.id });
 
-  res.status(200).json({
-    status: 'success',
-    message: 'deleted',
-  });
+  if (item.owner === req.user_id) {
+    await Item.deleteOne({ _id: req.params.id });
+
+    res.status(200).json({
+      status: 'success',
+      message: 'deleted',
+    });
+  } else {
+    res.status(403).json({
+      status: 'restricted',
+      message: 'you can not delete this item, because you are not its owner',
+    });
+  }
 };
 
 module.exports.getAllItems = async (req, res) => {
@@ -56,21 +66,29 @@ module.exports.getItem = async (req, res) => {
 };
 
 module.exports.updateItem = async (req, res) => {
-  const item = await Item.findOneAndUpdate(
-    { _id: req.params.id },
-    {
-      $set: {
-        title: req.body.title,
-        slug: req.body.slug,
-        price: req.body.price,
-        description: req.body.description,
-      },
-    },
-  );
+  const item = Item.findOne({ _id: req.params.id });
 
-  res.status(200).json({
-    status: 'success',
-    results: item.length,
-    data: { item },
-  });
+  if (item.owner === req.user_id) {
+    item.title = req.title;
+    item.slug = req.slug;
+    item.price = req.price;
+    item.description = req.description;
+
+    await item.save((err) => {
+      if (err) {
+        console.error(err);
+        console.log('error while saving');
+      }
+    });
+
+    res.status(200).json({
+      status: 'success',
+      message: 'item was saved successfully',
+    });
+  } else {
+    res.status(403).json({
+      status: 'restricted',
+      message: 'you can not delete this item, because you are not its owner',
+    });
+  }
 };
